ARM32 compilation fixes with GCC (dotnet/coreclr#27795)
* Suppress on clang only

* Fix integer conversion

* Extra qualifier

* Suppress warning

* Extra qualifier

* Signedness issues

* Correct offsetof

* Offsetof doesn't support non-constant values

* Conversion errors

* Move the comment too

* Fix assembly warning

* size is not constant

* Fix comment type

* Fix endmacro name

* Use OFFSET_NONE constant


Commit migrated from https://github.com/dotnet/coreclr/commit/e8bbcf18be67b864d16e12c13d16f1bea4a9c76e

+++ src/coreclr/src/ToolBox/superpmi/superpmi-shared/lightweightmap.h
@@ -240,7 +240,7 @@ public:
 
         // If we have RTTI, we can make this assert report the correct type. No RTTI, though, when
         // built with .NET Core, especially when built against the PAL.
-        AssertCodeMsg((ptr - rawData) == size, EXCEPTIONCODE_LWM, "%s - Ended with unexpected sizes %Ix != %x",
+        AssertCodeMsg((unsigned int)(ptr - rawData) == size, EXCEPTIONCODE_LWM, "%s - Ended with unexpected sizes %Ix != %x",
                       "Unknown type" /*typeid(_Item).name()*/, ptr - rawData, size);
     }
 
@@ -519,7 +519,7 @@ public:
             ptr += bufferLength * sizeof(unsigned char);
         }
 
-        AssertCodeMsg((ptr - rawData) == size, EXCEPTIONCODE_LWM, "Ended with unexpected sizes %Ix != %x",
+        AssertCodeMsg((unsigned int)(ptr - rawData) == size, EXCEPTIONCODE_LWM, "Ended with unexpected sizes %Ix != %x",
                       ptr - rawData, size);
     }
 
@@ -579,7 +579,7 @@ private:
             delete[] tItems;
         }
 
-        AssertCodeMsg((ptr - rawData) == size, EXCEPTIONCODE_LWM, "Ended with unexpected sizes %Ix != %x",
+        AssertCodeMsg((unsigned int)(ptr - rawData) == size, EXCEPTIONCODE_LWM, "Ended with unexpected sizes %Ix != %x",
                       ptr - rawData, size);
     }
